feat: badges are formatted for daterangepicker
